Make Kernel client test mocks order independent
